修复星尘代理拉起的进程如有子进程时无法全部结束的问题

+++ src/DH.SignalR/Dtos/OffLineData.cs
@@ -11,7 +11,7 @@ public class OffLineData
     /// <summary>
     /// 用户Id
     /// </summary>
-    public Int32 UserId { set; get; }
+    public Int64 UserId { set; get; }
 
     /// <summary>
     /// 连接Id
